requirements.txt
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
